float to double

+++ aluno.cpp
@@ -5,9 +5,9 @@
 #include "aluno.h"
 #define MAX 100
 
-void addAluno(tAlunoEntrada *vetAlunos, int &n, char *mod, int RA, float p1, float p2, float pt, float po)
+void addAluno(tAlunoEntrada *vetAlunos, int &n, char *mod, int RA, double p1, double p2, double pt, double po)
 {
-    float r1,r2,r3;
+    double r1,r2,r3;
 	strcpy(vetAlunos[n].nome, mod);
 	vetAlunos[n].RA = RA;
 	vetAlunos[n].p1 = p1;
@@ -64,7 +64,7 @@ for(int x = 0; x < n; x++){
 char* a = toLower(vetAlunosaux[x].nome);
 char* ab= toLower(nomedesejado);
 if(strstr(a,nomedesejado)){
-    printf("%s %.1f ",vetAlunos[x].nome,vetAlunos[x].media);
+    printf("%s %.1lf ",vetAlunos[x].nome,vetAlunos[x].media);
     if(vetAlunos[x].situacao==1){
     printf("Aprovado\n");
     }else{
@@ -81,10 +81,10 @@ if(strstr(a,nomedesejado)){
 }
 
 
-void optativa(float p1, float p2, float pt, float po, float* r1, float* r2, float* r3){
-    float temp = p1;
-    float maior = p2;
-    float trabalho = pt;
+void optativa(double p1, double p2, double pt, double po, double* r1, double* r2, double* r3){
+    double temp = p1;
+    double maior = p2;
+    double trabalho = pt;
     if(p1>p2){
         maior = p1;
         temp = p2;
@@ -97,8 +97,8 @@ void optativa(float p1, float p2, float pt, float po, float* r1, float* r2, floa
     *r3 = pt;
 }
 
-bool aprovador(float p1, float p2, float pt){
-float media = (0.35*p1)+(0.35*p2)+(0.3*pt);
+bool aprovador(double p1, double p2, double pt){
+double media = (0.35*p1)+(0.35*p2)+(0.3*pt);
 if(media>=6.0){
     return true;
 }else{
@@ -114,7 +114,7 @@ pont_aprov = fopen("Aprovados.txt", "w");
 
 for(int x = 0; x < n; x++)
     if(vetAlunos[x].situacao==1){
-    fprintf(pont_aprov,"%s %d %.1f\n",vetAlunos[x].nome,vetAlunos[x].RA,vetAlunos[x].media);
+    fprintf(pont_aprov,"%s %d %.1lf\n",vetAlunos[x].nome,vetAlunos[x].RA,vetAlunos[x].media);
     cont++;
 }
 fprintf(pont_aprov,"Total: %d",cont);
@@ -129,7 +129,7 @@ int cont = 0;
 pont_reprov = fopen("Reprovados.txt", "w");
 for(int x = 0; x < n; x++)
     if(vetAlunos[x].situacao==0){
-    fprintf(pont_reprov,"%s %d %.1f\n",vetAlunos[x].nome,vetAlunos[x].RA,vetAlunos[x].media);
+    fprintf(pont_reprov,"%s %d %.1lf\n",vetAlunos[x].nome,vetAlunos[x].RA,vetAlunos[x].media);
     cont++;
 }
 fprintf(pont_reprov,"Total: %d",cont);
@@ -140,7 +140,7 @@ void criaAtual(int &n, tAlunoEntrada *vetAlunos){
 FILE *pont_reprov;
 pont_reprov = fopen("Atual.txt", "w");
 for(int x = 0; x < n; x++)
-    fprintf(pont_reprov,"%s %d %.1f %.1f %.1f %.1f\n",vetAlunos[x].nome,vetAlunos[x].RA,vetAlunos[x].p1,vetAlunos[x].p2,vetAlunos[x].pt,vetAlunos[x].po);
+    fprintf(pont_reprov,"%s %d %.1lf %.1lf %.1lf %.1lf\n",vetAlunos[x].nome,vetAlunos[x].RA,vetAlunos[x].p1,vetAlunos[x].p2,vetAlunos[x].pt,vetAlunos[x].po);
 fclose(pont_reprov);
 }
 
@@ -159,7 +159,7 @@ if( arqEntrada == NULL ){
 
 for (int k = n; feof(arqEntrada)==0;k++){
     char tofix[50];
-    fscanf(arqEntrada, "%50[^0123456789] %d %f %f %f %f\n",tofix,&vetAlunos[k].RA,&vetAlunos[k].p1,&vetAlunos[k].p2,&vetAlunos[k].pt,&vetAlunos[k].po);
+    fscanf(arqEntrada, "%50[^0123456789] %d %lf %lf %lf %lf\n",tofix,&vetAlunos[k].RA,&vetAlunos[k].p1,&vetAlunos[k].p2,&vetAlunos[k].pt,&vetAlunos[k].po);
     tofix[strlen(tofix)-1] = '\0';
     strcpy(vetAlunos[k].nome,tofix);
     addAluno(vetAlunos,n,vetAlunos[k].nome,vetAlunos[k].RA,vetAlunos[k].p1,vetAlunos[k].p2,vetAlunos[k].pt,vetAlunos[k].po);
